agregando mas musica

+++ Proyecto_Alpha/Proyecto_Alpha/Form1.cs
@@ -23,9 +23,9 @@ namespace Proyecto_Alpha
         private SqlConnection conn;
         private string sCn;
         //Defino una variable de tipo Connection
-        private SqlConnection conn1;
+        //private SqlConnection conn1;
         //Defino una variable de tipo DataAdapter
-        private string sCn1;
+        //private string sCn1;
         private SqlCommand insert;
         //instacio un variable OleDbConection
         OleDbConnection cnn = new OleDbConnection();
@@ -46,12 +46,12 @@ namespace Proyecto_Alpha
             conn = new SqlConnection(sCn);
             //abro la conexión
             conn.Open();
-            cnn.ConnectionString = @"PROVIDER=SQLOLEDB;server= DESKTOP-F9FC1KG\\MSSQL; database= teatro ; Integrated Security=true;";
-            conexion cn1 = new conexion();
-            cn1.conec();
-            sCn1 = cn1.cadena;
-            conn1 = new SqlConnection(sCn1);
-            conn1.Open();
+            //cnn.ConnectionString = @"PROVIDER=SQLOLEDB;server= DESKTOP-F9FC1KG\\MSSQL; database= teatro ; Integrated Security=true;";
+            //conexion cn1 = new conexion();
+            //cn1.conec();
+            //sCn1 = cn1.cadena;
+            //conn1 = new SqlConnection(sCn1);
+            //conn1.Open();
 
 
 
@@ -63,7 +63,7 @@ namespace Proyecto_Alpha
             n.Show();
         }
         // cambiar la direccion por cada computadora
-
+       
         Image Rojo = Image.FromFile(@"C:\Users\c-arm\Documents\Visual Studio 2019\StartPages\Proyecto_Alpha\Proyecto_Alpha\Silla_Roja.png");
         Image Negro = Image.FromFile(@"C:\Users\c-arm\Documents\Visual Studio 2019\StartPages\Proyecto_Alpha\Proyecto_Alpha\Silla_Negra.png");
 
@@ -71,7 +71,7 @@ namespace Proyecto_Alpha
         {
             string s;
             DateTime fech;
-            if (cmbHora.SelectedItem == null) { }
+            if (cmbHora.SelectedItem == null) { MessageBox.Show("Debe seleccionar hora"); }
             else
             {
                  s = cmbHora.SelectedItem.ToString();
@@ -92,66 +92,68 @@ namespace Proyecto_Alpha
             }
 
             String fullimagepath = Path.Combine(Application.StartupPath, "Archivo\\" + cmbFuncion.Text + ".txt");
-            
+
 
 
             // MessageBox.Show(fullimagepath);
-
-            if (File.Exists(fullimagepath))
-            {
-
-              //  StreamReader sr = new StreamReader(fullimagepath, true);
-                StreamWriter sw = new StreamWriter(fullimagepath, true);
-
-
-
-                sw.WriteLine("Bienvenido al cine SOYACITY");
-                sw.WriteLine("\nLa cantidad de Asientos comprados son: " + lslButaca.Items.Count);
-                sw.WriteLine("\nLos asientos comprados son");
-
-
-
-                foreach (ListViewItem item in lslButaca.Items)
+            if (lslButaca.Items.Count==0) { MessageBox.Show("Debe Comprar asientos"); }
+            else {
+                if (File.Exists(fullimagepath))
                 {
-                    sw.WriteLine(" " + item.Text);
+
+                    //  StreamReader sr = new StreamReader(fullimagepath, true);
+                    StreamWriter sw = new StreamWriter(fullimagepath, true);
+
+
+
+                    sw.WriteLine("Bienvenido al cine SOYACITY");
+                    sw.WriteLine("\nLa cantidad de Asientos comprados son: " + lslButaca.Items.Count);
+                    sw.WriteLine("\nLos asientos comprados son");
+
+
+
+                    foreach (ListViewItem item in lslButaca.Items)
+                    {
+                        sw.WriteLine(" " + item.Text);
+
+
+                    }
+                    sw.WriteLine("Fechan y hora: " + cmbHora.Text);
+
+                    sw.WriteLine(" Total: " + lblMoney.Text + " $");
+                    sw.WriteLine("---------------------------------------------------------------------------");
+                    sw.Close();
+
+                }
+                else
+                {
+                    // StreamReader sr = new StreamReader(fullimagepath, true);
+                    StreamWriter sw = new StreamWriter(fullimagepath, true);
+
+                    sw.WriteLine("Bienvenido al cine SOYACITY");
+                    sw.WriteLine("\nLa cantidad de Asientos comprados son: " + lslButaca.Items.Count);
+                    sw.WriteLine("\nLos asientos comprados son");
+
+
+
+                    foreach (ListViewItem item in lslButaca.Items)
+                    {
+                        sw.WriteLine(" " + item.Text);
+
+
+                    }
+                    sw.WriteLine("Fechan y hora: ");
+
+                    sw.WriteLine(" Total: " + lblMoney.Text + " $");
+                    sw.WriteLine("---------------------------------------------------------------------------");
+
+
+                    sw.Close();
 
 
                 }
-                sw.WriteLine("Fechan y hora: ");
-
-                sw.WriteLine(" Total: " + lblMoney.Text + " $");
-                sw.WriteLine("---------------------------------------------------------------------------");
-                sw.Close();
-
+                Process.Start(fullimagepath);
             }
-            else
-            {
-               // StreamReader sr = new StreamReader(fullimagepath, true);
-                StreamWriter sw = new StreamWriter(fullimagepath, true);
-
-                sw.WriteLine("Bienvenido al cine SOYACITY");
-                sw.WriteLine("\nLa cantidad de Asientos comprados son: " + lslButaca.Items.Count);
-                sw.WriteLine("\nLos asientos comprados son");
-
-
-
-                foreach (ListViewItem item in lslButaca.Items)
-                {
-                    sw.WriteLine(" " + item.Text);
-
-
-                }
-                sw.WriteLine("Fechan y hora: ");
-
-                sw.WriteLine(" Total: " + lblMoney.Text + " $");
-                sw.WriteLine("---------------------------------------------------------------------------");
-
-
-                sw.Close();
-
-
-            }
-            Process.Start(fullimagepath);
             musica.controls.stop();
         }
 
@@ -271,9 +273,16 @@ namespace Proyecto_Alpha
         private void cmbFuncion_SelectedIndexChanged(object sender, EventArgs e)
         {
 
-            //MessageBox.Show(ruta);
-            musica.URL = ruta;
-            musica.controls.play();
+           // MessageBox.Show(ruta);
+            if (ruta == null) {
+                cmbFuncion.Text = " ";
+            }
+            else
+            {
+                musica.URL = ruta;
+                musica.controls.play();
+            }
+            
             hora.Start();
 
            
@@ -283,7 +292,7 @@ namespace Proyecto_Alpha
         private void Form1_Load(object sender, EventArgs e)
         {
             tipo.Text = login.tipo;
-            if (tipo.Text == "Admin")
+            if (tipo.Text == " Admin")
             {
                 menuToolStripMenuItem.Visible = true;
                 btnEliminarF.Visible = true;
@@ -2298,6 +2307,17 @@ namespace Proyecto_Alpha
             }
             hora.Start();
             cmbFuncion.SelectedItem = null;
+        }
+
+        private void MenuStrip1_ItemClicked(object sender, ToolStripItemClickedEventArgs e)
+        {
+
+        }
+
+        private void AgregarPeliculaToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            Agregar_Evento n = new Agregar_Evento();
+            n.Show();
         }
     }
 }
